requirements.txt
google-api-core==2.17.1
google-cloud-documentai==2.24.1